Implement level up language flow
