+++ frontend/src/components/admin/ApprovedLanguagesTable.tsx
@@ -1,22 +1,7 @@
-import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
+import React, { useState } from "react";
 import { useMutation } from "@apollo/client";
-import {
-  Table,
-  Thead,
-  Tbody,
-  Tooltip,
-  Tr,
-  Th,
-  Td,
-  Tfoot,
-  Button,
-  Slider,
-  SliderTrack,
-  SliderFilledTrack,
-  SliderThumb,
-  SliderMark,
-} from "@chakra-ui/react";
-import { ApprovedLanguagesMap, generateSortFn } from "../../utils/Utils";
+import { useHistory } from "react-router-dom";
+import { ApprovedLanguagesMap } from "../../utils/Utils";
 import { convertLanguageTitleCase } from "../../utils/LanguageUtils";
 import {
   UPDATE_USER_APPROVED_LANGUAGES,
@@ -25,21 +10,20 @@ import {
 import ApproveLanguageModal, {
   NewApprovedLanguage,
 } from "../utils/ApproveLanguageModal";
+import {
+  CREATE_TRANSLATION_TEST,
+  CreateTranslationTestResponse,
+} from "../../APIClients/mutations/StoryMutations";
+import ApprovedLanguagesTableComponent from "./ApprovedLanguagesTableComponent";
+import ConfirmationModal from "../utils/ConfirmationModal";
 
-interface ApprovedLanguageFieldSortDict {
-  [field: string]: {
-    sortFn: (a: ApprovedLanguage, b: ApprovedLanguage) => number;
-    isAscending: boolean;
-    setIsAscending: Dispatch<SetStateAction<boolean>>;
-  };
-}
-
-type ApprovedLanguage = {
-  language: string;
-  level: number;
-  role: string;
-  sliderValue: number;
-};
+import {
+  LEVEL_UP_LANGUAGE_CONFIRMATION,
+  LEVEL_UP_LANGUAGE_BUTTON,
+  NEW_BOOK_TEST_ADDED_HEADING,
+  NEW_BOOK_TEST_ADDED_MESSAGE,
+  NEW_BOOK_TEST_ADDED_BUTTON,
+} from "../../utils/Copy";
 
 export type ApprovedLanguagesTableProps = {
   approvedLanguagesTranslation: ApprovedLanguagesMap | undefined;
@@ -68,12 +52,12 @@ const ApprovedLanguagesTable = ({
   setAlertTimeout,
   isAdmin = false,
 }: ApprovedLanguagesTableProps) => {
-  const [approvedLanguages, setApprovedLanguages] = useState<
-    ApprovedLanguage[]
-  >([]);
+  const history = useHistory();
 
-  const [isAscendingLanguage, setIsAscendingLanguage] = useState(true);
-  const [isAscendingRole, setIsAscendingRole] = useState(true);
+  const [confirmLevelUp, setConfirmLevelUp] = useState(false);
+  const [alertNewTest, setAlertNewTest] = useState(false);
+  const [testLevel, setTestLevel] = useState(0);
+  const [testLanguage, setTestLanguage] = useState("");
 
   const [approveNewLanguage, setApproveNewLanguage] = useState(false);
 
@@ -99,6 +83,9 @@ const ApprovedLanguagesTable = ({
   const [updateUserApprovedLanguages] = useMutation<{
     response: UpdateUserApprovedLanguagesResponse;
   }>(UPDATE_USER_APPROVED_LANGUAGES);
+  const [createTranslationTest] = useMutation<{
+    response: CreateTranslationTestResponse;
+  }>(CREATE_TRANSLATION_TEST);
 
   const getAlertText = (
     isTranslate: boolean,
@@ -157,45 +144,6 @@ const ApprovedLanguagesTable = ({
     }
   };
 
-  const fieldSortDict: ApprovedLanguageFieldSortDict = {
-    language: {
-      sortFn: generateSortFn<ApprovedLanguage>("language", isAscendingLanguage),
-      isAscending: isAscendingLanguage,
-      setIsAscending: setIsAscendingLanguage,
-    },
-    role: {
-      sortFn: generateSortFn<ApprovedLanguage>("role", isAscendingRole),
-      isAscending: isAscendingRole,
-      setIsAscending: setIsAscendingRole,
-    },
-  };
-
-  const sortApprovedLanguages = (field: string) => {
-    const newApprovedLanguages = [...approvedLanguages];
-    const { sortFn, isAscending, setIsAscending } = fieldSortDict[field];
-    setIsAscending(!isAscending);
-    newApprovedLanguages.sort(sortFn);
-    setApprovedLanguages(newApprovedLanguages);
-  };
-
-  useEffect(() => {
-    if (approvedLanguagesTranslation && approvedLanguagesReview) {
-      const translationLanguages = Object.entries(
-        approvedLanguagesTranslation,
-      ).map(([language, level]) => {
-        return { language, level, role: "Translator", sliderValue: level };
-      });
-      const reviewLanguages = Object.entries(approvedLanguagesReview).map(
-        ([language, level]) => {
-          return { language, level, role: "Reviewer", sliderValue: level };
-        },
-      );
-      const combined = translationLanguages.concat(reviewLanguages);
-      sortApprovedLanguages("language");
-      setApprovedLanguages(combined);
-    }
-  }, [approvedLanguagesTranslation, approvedLanguagesReview]);
-
   const onSliderValueChange = (
     isTranslate: boolean,
     language: string,
@@ -221,122 +169,50 @@ const ApprovedLanguagesTable = ({
     closeApproveLanguageModal(false);
   };
 
-  const tableBody = approvedLanguages.map(
-    (approvedLanguage: ApprovedLanguage, index: number) => (
-      <Tr
-        borderBottom={
-          index === approvedLanguages.length - 1 ? "1px solid gray" : ""
-        }
-        key={`${approvedLanguage.language}/${approvedLanguage.role}`}
-      >
-        <Td>{convertLanguageTitleCase(approvedLanguage.language)}</Td>
-        <Td>{approvedLanguage.role}</Td>
-        <Td colSpan={4} align="center">
-          <Slider
-            isDisabled={!isAdmin}
-            defaultValue={approvedLanguage.level}
-            min={1}
-            max={4}
-            step={1}
-            marginLeft="20px"
-            size="lg"
-            width="79%"
-            onChange={(val: number) =>
-              onSliderValueChange(
-                approvedLanguage.role === "Translator",
-                approvedLanguage.language,
-                val,
-                approvedLanguage.level,
-              )
-            }
-          >
-            <SliderTrack height="6px">
-              <SliderFilledTrack bg="blue.500" />
-            </SliderTrack>
-            <SliderThumb boxShadow="0px 0px 3px black" />
-            {[1, 2, 3, 4].map((val) => (
-              <SliderMark
-                key={val}
-                value={val}
-                height="6px"
-                width="4px"
-                marginTop="-3px"
-                // #64A1CD is light blue, #ECF1F4 is light gray
-                bg={approvedLanguage.sliderValue > val ? "#64A1CD" : "#ECF1F4"}
-              />
-            ))}
-          </Slider>
-        </Td>
-        {!isAdmin && (
-          <Tooltip
-            hasArrow
-            label="Currently at maximum language level."
-            isDisabled={approvedLanguage.level !== 4}
-          >
-            <Td>
-              <Button
-                aria-label="Level up approval language and level"
-                background="white"
-                size="sm"
-                width="103px"
-                variant="blueOutline"
-                isDisabled={approvedLanguage.level === 4}
-              >
-                Level Up
-              </Button>
-            </Td>
-          </Tooltip>
-        )}
-      </Tr>
-    ),
-  );
+  const closeNewTestModal = () => {
+    setAlertNewTest(false);
+  };
+
+  const closeLevelUpModal = () => {
+    setTestLevel(0);
+    setTestLanguage("");
+    setConfirmLevelUp(false);
+  };
+
+  const openLevelUpModal = (level: number, language: string) => {
+    setTestLevel(level);
+    setTestLanguage(language);
+    setConfirmLevelUp(true);
+  };
+
+  const callCreateTranslationTestMutation = async () => {
+    try {
+      await createTranslationTest({
+        variables: {
+          userId,
+          level: testLevel,
+          language: testLanguage,
+        },
+      });
+      closeLevelUpModal();
+      setAlertNewTest(true);
+    } catch (err) {
+      window.alert(err);
+      closeLevelUpModal();
+      closeNewTestModal();
+    }
+  };
 
   return (
     <>
-      <Table
-        borderRadius="12px"
-        boxShadow="0px 0px 2px grey"
-        theme="gray"
-        variant="striped"
-        width="100%"
-        marginTop="20px"
-      >
-        <Thead>
-          <Tr
-            borderTop="1em solid transparent"
-            borderBottom="0.5em solid transparent"
-          >
-            <Th
-              cursor="pointer"
-              onClick={() => sortApprovedLanguages("language")}
-              width="12%"
-            >
-              {`LANGUAGE ${isAscendingLanguage ? "↑" : "↓"}`}
-            </Th>
-            <Th
-              cursor="pointer"
-              onClick={() => sortApprovedLanguages("role")}
-              width="12%"
-            >{`ROLE ${isAscendingRole ? "↑" : "↓"}`}</Th>
-            {["LEVEL 1", "LEVEL 2", "LEVEL 3", "LEVEL 4"].map((level) => (
-              <Th key={level}>{level}</Th>
-            ))}
-            {!isAdmin && <Th width="7%">ACTION</Th>}
-          </Tr>
-        </Thead>
-        <Tbody>{tableBody}</Tbody>
-        <Tfoot padding="12px">
-          <Button
-            margin="16px 0px 16px 12px"
-            size="secondary"
-            variant="blueOutline"
-            width="254px"
-            onClick={openApproveLanguageModal}
-          >
-            Add New Language
-          </Button>
-        </Tfoot>
-      </Table>
+      <ApprovedLanguagesTableComponent
+        addNewLanguage={isAdmin ? openApproveLanguageModal : () => undefined}
+        levelUpOnClick={openLevelUpModal}
+        onSliderValueChange={onSliderValueChange}
+        approvedLanguagesTranslation={approvedLanguagesTranslation}
+        approvedLanguagesReview={approvedLanguagesReview}
+        isAdmin={isAdmin}
+      />
       {approveNewLanguage && (
         <ApproveLanguageModal
           isOpen={approveNewLanguage}
@@ -346,6 +222,21 @@ const ApprovedLanguagesTable = ({
           approvedLanguagesReview={approvedLanguagesReview!}
         />
       )}
+      <ConfirmationModal
+        confirmation={confirmLevelUp}
+        onClose={closeLevelUpModal}
+        onConfirmationClick={callCreateTranslationTestMutation}
+        confirmationMessage={LEVEL_UP_LANGUAGE_CONFIRMATION}
+        buttonMessage={LEVEL_UP_LANGUAGE_BUTTON}
+      />
+      <ConfirmationModal
+        confirmation={alertNewTest}
+        onClose={closeNewTestModal}
+        onConfirmationClick={() => history.push("/")}
+        confirmationHeading={NEW_BOOK_TEST_ADDED_HEADING}
+        confirmationMessage={NEW_BOOK_TEST_ADDED_MESSAGE}
+        buttonMessage={NEW_BOOK_TEST_ADDED_BUTTON}
+      />
     </>
   );
 };
